Refactor ChooseConfiguration component: extract configuration details display into a new ConfigTable component, improving code organization and readability.
